Replace more custom types

+++ Sources/Engine/Light/LensFlares.h
@@ -39,7 +39,7 @@ public:
   float olf_fLightDesaturation;       // desaturation factor for light color (1=monochrome)
   float oft_fFallOffFactor;           // flare falloff relative to light falloff
   ANGLE olf_aRotationFactor;          // flare rotation (deg/screen width)
-  ULONG olf_ulFlags;
+  uint32_t olf_ulFlags;
 };
 
 /*
